requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.2.1=pypi_0
backcall=0.2.0=pypi_0
bcrypt=4.0.1=pypi_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
cmake=3.26.1=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.7=pypi_0
cryptography=40.0.1=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=pypi_0
executing=1.2.0=pypi_0
filelock=3.10.7=pypi_0
fonttools=4.39.4=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.1.0=pypi_0
importlib-resources=5.12.0=pypi_0
ipdb=0.13.13=pypi_0
ipython=8.12.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lit=16.0.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
ml-collections=0.1.1=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.0=pypi_0
numpy=1.24.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
openssl=1.1.1t=h7f8727e_0
packaging=23.1=pypi_0
pandas=1.5.3=pypi_0
paramiko=3.1.0=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=23.0.1=py39h06a4308_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.22.1=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pypi_0
pygments=2.15.1=pypi_0
pynacl=1.5.0=pypi_0
pyparsing=3.0.9=pypi_0
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
scipy=1.10.1=pypi_0
sentry-sdk=1.18.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py39h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.41.1=h5eee18b_0
stack-data=0.6.2=pypi_0
sumolib=1.16.0=pypi_0
sympy=1.11.1=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=2.0.0=pypi_0
torchaudio=2.0.1=pypi_0
torchvision=0.15.1=pypi_0
tqdm=4.65.0=pypi_0
traci=1.16.0=pypi_0
traitlets=5.9.0=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.15=pypi_0
wandb=0.14.0=pypi_0
wcwidth=0.2.6=pypi_0
wheel=0.38.4=py39h06a4308_0
xz=5.2.10=h5eee18b_1
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
